[#1265] MVC - Minimum Viable Compilation
Minium changes required to attempt authentication with Keycloak.  In
the past we have tried using the Servlet Filter adapter [1] but this
time we're using the Spring Security adapter [2]

This code expects to have a client defined as `akvo-flow` with an
installation file (keycloak.json) under GAE/war/WEB-INF/keycloak.json

We also require to have a test user `akvo.flow.user.test@gmail.com`
created as a user in Keycloak.

The current setup of Roles in Akvo Flow implies a hierarchy:

    SUPER_ADMIN <-- ADMIN <-- USER

We created the role `USER` in Keycloak under the `akvo-flow` client
and associate it with the test user.

It seems that we need to fiddle to role mappings if we want to keep
the expressions in webapp-security.xml, e.g. `hasRole('USER')`

An unsolved problem is how to map the instance userId to the principal
(uuid) in Keycloak.

The BaseDAO uses the principal as the userId in the datastore, but we don't
have that information in the SecurityContext.

We could subclass Keycloak implementation and wrap
`KeycloakAuthenticationToken` adding information about the userId, but
this leave us with the task of maintaining and updating this code
(e.g. when we upgrade libraries)

Another approach is to extend the `User` kind in Flow, adding another
property that is the uuid in Keycloak.

The main problem is that we don't want to add extra dependencies and
extra work to the `BaseDAO` class (e.g. Query the datastore every time
we need to filter data)

[1] https://www.keycloak.org/docs/latest/securing_apps/index.html#_servlet_filter_adapter
[2] https://www.keycloak.org/docs/latest/securing_apps/index.html#_spring_security_adapter